test(pii): consolidate analyzer and anonymizer integration tests
Merge analyzer_behaviour, analyzer_builder, and analyzer_corpus into a
single tests/analyzer.rs, and merge anonymizer_acceptance,
anonymizer_roundtrip, and hash_determinism into tests/anonymizer.rs.
Collapses six integration-test binaries into two and shares helper
fixtures within each file.
